Problem: #69 App architecture inefficiencies and overhead
Solution: moved more of ui to imports. #69
