+++ imports/ui/components/bounties/HashrateAPI.html
@@ -1,0 +1,41 @@
+<template name="HashrateAPI">
+  <h5>API Bounty for {{thisbounty.currencyName}}</h5>
+  <p><b>Instructions:</b> find a block explorer or an open node for {{currencyName}} which has an API.
+    Look at the API's documentation to find out how to construct an API call. It's possible (but unlikely) that you may need to register for an API key and include it in the call.
+    Construct the API call in the box below, and examine the response. <b>Hint:</b> open your browser's javascript console for diagnostic information.
+    The result is returned as a JSON object called "data". Please navigate to the correct place in the JSON object using dot notation and only use square brackets if dot notation is not possible.
+  </p>
+  <p>If you cannot find an API providing network hashpower directly, you may need to use difficulty instead and work out what to do with the difficulty number in order to get the hashpower (you can use standard javascript math operators).</p>
+  <p>Do <b>not</b> use an API from a third party site like coinwarz or bitinfocarts, you must use a native block explorer or an open node</p>
+  <p><b>The form has been pre-filled with a working example for Monero to help get you started.</b></p>
+  <div style="display:inline-flex;">fetch(<input id="apiCall" type="text" style="width:600px;" placeholder="Add the entire API call here starting with http..." value="https://moneroblocks.info/api/get_stats">)</div><br />
+  .then<span style="color: blue">(</span>function(result) { <br />
+  &nbsp;&nbsp;&nbsp;&nbsp;return result.json(); <br />
+  }<span style="color: blue">)</span>.then<span style="color: blue">(</span>function(<span style="color: red">data</span>) { <br />
+  console.log(Math.round(<input id="hashrate" type="text" style="width:300px;" value="data.hashrate">)); // in {{hashpowerunits}}/s.<br />
+  }<span style="color: blue">)</span>; <br />
+  <b>Sometimes the same API might have some other data:</b> <br />
+  Please also fill these out if they exist. Make them blank if the JSON data doesn't exist.<br />
+  <ul>
+    <li>Blockchain Height: <input id="height" type="text" style="width:300px;" value="data.height"> </li>
+    <li>Timestamp: <input id="timestamp" type="text" style="width:300px;" value="data.last_timestamp"> </li>
+    <li>Block reward: <input id="reward" type="text" style="width:300px;" value="data.last_reward"> </li>
+    <li>Total coins in existence: <input id="emission" type="text" style="width:300px;" value="data.total_emission"> </li>
+  </ul>
+  <b>Most currencies report coins in 'atomic units'</b>, e.g. Bitcoin uses the Satoshi. Look at any coin counts (e.g. the block reward) in the JSON data below and compare this to what is reported on the block explorer website to see if they move the decimal point.
+  Move decimal this many places to the left: <input id="decimal" type="text" style="width:30px;" value="12"> <br />
+  <button class="btn btn-warning fetch">Execute API Call</button> <br />
+  <b>Parsed Response (should look like a number):</b>
+  <div>
+    <ul>
+      <li>Hashrate: <span id="hashrate-result"></span> </li>
+      <li>Blockchain Height: <span id="height-result"></span> </li>
+      <li>Timestamp: <span id="timestamp-result"></span> </li>
+      <li>Block reward: <span id="reward-result"></span> </li>
+      <li>Total coins in existence: <span id="emission-result"></span> </li>
+    </ul>
+  <button class="btn btn-primary submit">This looks good!</button>
+  </div>
+  <h6 style="margin-top: 10px;">Full JSON response:</h6>
+  <div id="result">JSON data will display here</div>
+  </template>
