+++ imports/ui/components/bounties/bountyRender.html
@@ -1,0 +1,37 @@
+<template name="bountyRender">
+  <div class="card my-3">
+    <div class="card-body">
+      <div class="row">
+        <div class="col-sm-8">
+          <h5 class="card-title mb-3">{{types.heading}} for {{currencyName}}</h5>
+          <h6 class="card-subtitle mb-2 text-muted"></h6>
+
+          <div class="card-text"><b>Problem:</b> {{problem}}.</div>
+          <div class="card-text"><b>Solution:</b> {{solution}}.</div>
+          <div class="card-text" id={{id}} style="display:none;margin-top:10px;">
+            <h6><b>Bounty Details</b></h6>
+            {{#if canContinue}}<b>(In progress)</b>{{/if}} {{types.rules}}
+
+            <div class="mt-3">
+              <button type="button" class="btn btn-danger cancel">Cancel</button>
+              <button type="button" class="btn btn-primary start" style="margin-left:10px;">{{#if canContinue}}Continue{{else}}Start{{/if}}</button>
+            </div>
+          </div>
+        </div>
+        <div class="col-sm-4">
+          <h5 class="float-right my-4 my-sm-0">
+            <span class="mr-2">{{reward}}</span>KZR
+          </h5>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col">
+          <div class="float-right" style="display:inline-flex;clear:both;" id="takeBounty{{id}}">
+            <button type="button" class="btn {{buttonClass}}">{{#if canContinue}}Continue{{else}}Take{{/if}} Bounty</button>
+          </div>
+          <div class="float-right" style="display:inline-flex;clear:both;"><i>{{workingText}}</i></div>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
