+++ imports/ui/components/bounties/miscellaneous.html
@@ -1,0 +1,19 @@
+<template name="hashpower">
+  hashpower
+  </template>
+  
+  
+  <template name="userExperience">
+  userExperience
+  </template>
+  
+  
+  <template name="codeQuality">
+  codeQuality
+  </template>
+  
+  
+  <template name="community">
+  community
+  </template>
+
