+++ imports/ui/components/bounties/miscellaneous.js
@@ -1,0 +1,3 @@
+import { Template } from 'meteor/templating';
+
+import './miscellaneous.html'
