+++ imports/ui/components/bounties/rainbow.js
