+++ imports/ui/pages/addCoin.html
