+++ imports/ui/pages/bounties.html
@@ -1,0 +1,18 @@
+<template name="bounties">
+  <!-- {{#unless reactiveVar 'workingBounty'}} -->
+  <h3 class="mt-3">Bounties</h3>
+  {{#if subsCacheReady}}
+    {{#each bounties}}
+      {{> bountyRender}}
+    {{else}}
+      {{> empty}}
+    {{/each}}
+  {{else}}
+    {{> loading}}
+  {{/if}}
+  <!-- {{/unless}}
+
+  {{#if reactiveVar 'workingBounty'}}
+  {{#each activebounty}}{{> activeBounty}}{{/each}}
+  {{/if}} -->
+</template>
