+++ imports/ui/pages/communities.js
@@ -1,0 +1,228 @@
+import { Template } from 'meteor/templating';
+import { Currencies, Ratings, Bounties } from '/imports/api/indexDB.js';
+import Cookies from 'js-cookie'
+
+import './communities.html'
+
+Template.communities.onCreated(function() {
+    this.autorun(() => {
+        SubsCache.subscribe('approvedcurrencies')
+        SubsCache.subscribe('ratings')
+        SubsCache.subscribe('communityBounty')
+    })
+
+    this.name = new ReactiveVar('')
+    this.symbol = new ReactiveVar('')
+    this.cnt = 0
+    this.ties = 0
+
+    this.now = new ReactiveVar(Date.now())
+    Meteor.setInterval(() => {
+        this.now.set(Date.now())
+    }, 1000)
+})
+
+Template.communities.onRendered(function() {
+    this.autorun(function() {
+        if (Template.instance().subscriptionsReady()) {
+            const count = Ratings.find({
+                $or: [{
+                    answered: false,
+                    catagory: 'community'
+                }, {
+                    answered: false,
+                    context: 'community'
+                }]
+            }).count()
+
+            $('#outstandingRatings').toggle(!!count)
+            $('#currencyChoices').toggle(!count)
+        }
+    })
+})
+
+Template.communities.events({
+    'click #elo': (event, templateInstance) => {
+        Meteor.call('tabulateElo', (err, data) => {})
+    },
+    'click #communities': (event, templateInstance) => {
+        Meteor.call('averageElo', 'community', (err, data) => {})
+    },
+    'keyup #js-name, keyup #js-symbol': (event, templateInstance) => {
+        event.preventDefault()
+
+        templateInstance[event.currentTarget.id.substring(3)].set($(event.currentTarget).val())
+    },
+    'click #populateRatings': (event, templateInstance) => {
+        Meteor.call('populateCommunityRatings', (err, result) => {
+            if (err) {
+                sAlert.error(err.reason)
+            } else {
+                if (!Ratings.findOne({
+                    $or: [{
+                        answered: false,
+                        catagory: 'community'
+                    }, {
+                        answered: false,
+                        context: 'community'
+                    }]
+                })) {
+                    sAlert.error('Please add some communities to continue.')
+                }
+            }
+        })
+    },
+    'mouseover .choice': (event, templateInstance) => {
+        $('.choice').css('cursor', 'pointer')
+    },
+    'click #js-cancel': (event, templateInstance) => {
+        event.preventDefault()
+
+        Meteor.call('deleteNewBountyClient', 'new-community', (err, data) => {})
+        Cookies.set('workingBounty', false, { expires: 1 })
+
+        FlowRouter.go('/')
+    },
+    'click .choice': function(event, templateInstance) {
+        if (event.currentTarget.id === 'tie') {
+            templateInstance.ties++
+        } else {
+            templateInstance.ties = 0
+        }
+
+        Meteor.call('answerCommunityRating', this._id, event.currentTarget.id, (err, data) => {
+            if (err && err.reason === 'xor') {
+                if (templateInstance.cnt++ === 0) {
+                    sAlert.error('Your answer is in contradiction with your previous answers. Please try again. If this persists, your progress will be purged and bounties will be nullified.')
+                } else {
+                    sAlert.error('Lazy answering detected. You\'ll have to start all over again.')
+                    Meteor.call('deleteCommunityRatings', (err, data) => {})
+
+                    templateInstance.cnt = 0
+                }
+            }
+
+            Cookies.set('workingBounty', false, { expires: 1 })
+
+            if (templateInstance.ties > 10) { // ties can't be checked with XOR questions, as XOR only works on booleans. Nonetheless, if the user clicks on 'tie' 10 times in a row, it's safe to say that he/she is just lazy answering
+                sAlert.error('Lazy answering detected. You\'ll have to start all over again.')
+                Meteor.call('deleteCommunityRatings', (err, data) => {})
+
+                templateInstance.ties = 0
+            }
+        })
+    },
+    'click .js-save': function(event, templateInstance) {
+        Meteor.call('saveCommunity', this._id, $(`#js-com-url_${this._id}`).val(), (err, data) => {
+            if (!err) {
+                $(`#js-com-url_${this._id}`).attr('disabled', 'true')
+                $(event.currentTarget).attr('disabled', 'true')
+                $(event.currentTarget).text('Saved.')
+
+                setTimeout(() => $(`#links_${this._id}`).hide(), 1000)
+            } else {
+                sAlert.error(err.reason)
+            }
+        })
+    },
+    'click .toggle': function(event, templateInstance) {
+        $(`#links_${this._id}`).toggle()
+    }
+})
+
+Template.communities.helpers({
+    activeBounty: () => {
+        let bounty = Bounties.find({
+            userId: Meteor.userId(),
+            type: 'new-community',
+            completed: false
+        }, {
+            sort: {
+                expiresAt: -1
+            }
+        }).fetch()[0]
+
+        return bounty && bounty.expiresAt > Date.now()
+    },
+    timeRemaining: () => {
+        let bounty = Bounties.find({
+            userId: Meteor.userId(),
+            type: 'new-community',
+            completed: false
+        }, {
+            sort: {
+                expiresAt: -1
+            }
+        }).fetch()[0]
+      
+        return `You have ${Math.round((bounty.expiresAt - Template.instance().now.get())/1000/60)} minutes to complete the bounty for ${Number(bounty.currentReward).toFixed(2)} (* number of questions) KZR.`;
+    },
+    outstandingRatings: () => {
+        return Ratings.find({
+            $or: [{
+                answered: false,
+                catagory: 'community'
+            }, {
+                answered: false,
+                context: 'community'
+            }]
+        }).count()
+    },
+    alreadyAdded: () => {
+        let alreadyAdded = _.uniq(_.flatten(Ratings.find({
+            $or: [{
+                owner: Meteor.userId(),
+                catagory: 'community'
+            }, {
+                owner: Meteor.userId(),
+                context: 'community'
+            }]
+        }).fetch().map(i => [i.currency0Id,i.currency1Id])))
+
+        return Currencies.find({
+            _id: {
+                $in: alreadyAdded
+            }
+        })
+    },
+    currencies: () => {
+        let alreadyAdded = _.uniq(_.flatten(Ratings.find({
+            $or: [{
+                owner: Meteor.userId(),
+                catagory: 'community'
+            }, {
+                owner: Meteor.userId(),
+                context: 'community'
+            }]
+        }).fetch().map(i => [i.currency0Id,i.currency1Id])))
+
+        return Currencies.find({
+            _id: {
+                $nin: alreadyAdded
+            },
+            currencyName: new RegExp(Template.instance().name.get(), 'ig'),
+            currencySymbol: new RegExp(Template.instance().symbol.get(), 'ig')
+        })
+    },
+    questions: () => {
+        return Ratings.findOne({
+            $or: [{
+                answered: false,
+                catagory: 'community'
+            }, {
+                answered: false,
+                context: 'community'
+            }]
+        })
+    },
+    currency0Name: function() {
+        return Currencies.findOne({
+            _id: this.currency0Id
+        }).currencyName
+    },
+    currency1Name: function() {
+        return Currencies.findOne({
+            _id: this.currency1Id
+        }).currencyName
+    }
+})
