+++ imports/ui/pages/editProfile.html
@@ -1,4 +1,4 @@
-<template name="viewprofile">
+<template name="editProfile">
     <h3>Edit Profile</h3>
     <form id="editProfile">
         <div class="form-group">
