+++ imports/ui/pages/flaggedUsers/flaggedUsers.html
@@ -1,0 +1,35 @@
+<template name="flaggedUsers">
+    <h3>Flagged users</h3>
+    {{#if isModerator}}
+    <table class="table">
+        <thead>
+            <tr>
+                <th scope="col">Username</th>
+                <th scope="col">Email</th>
+                <th scope="col">IP address</th>
+                <th scope="col">Last access</th>
+                <th scope="col">Active flags</th>
+            </tr>
+        </thead>
+        <tbody>
+            {{#if subsCacheReady}}
+                {{#each users}}
+                    <tr>
+                        <td>{{user.username}}</td>
+                        <td>{{user.email}}</td>
+                        <td>{{ipAddress}}</td>
+                        <td>{{lastAccess}}</td>
+                        <td>{{activeFlags}}</td>
+                    </tr>
+                {{else}}
+                    {{> empty}}
+                {{/each}}
+            {{else}}
+                {{> loading}}
+            {{/if}}
+        </tbody>
+    </table>
+    {{else}}
+    <h5>You have to be a moderator to view this page.</h5>
+    {{/if}}
+</template>
