+++ imports/ui/pages/hashpower/allHashaverage.html
@@ -1,0 +1,26 @@
+<template name="allHashaverage">
+    <h3>Hash power averages</h3>
+    <table class="table">
+        <thead>
+            <tr>
+                <th scope="col">Algorithm</th>
+                <th scope="col">Average power consumption per H/s (sol/s)</th>
+            </tr>
+        </thead>
+        <tbody>
+            {{#if subsCacheReady}}
+                {{#each hashaverage}}
+                    <tr>
+                        <td>{{hashAlgorithm}}</td>
+                        <td>{{average}}</td>
+                    </tr>
+                {{else}}
+                    {{> empty}}
+                {{/each}}
+            {{else}}
+                {{> loading}}
+            {{/if}}
+        </tbody>
+    </table>
+    <button class="btn btn-danger" id="js-update">Update averages</button>
+</template>
