+++ imports/ui/pages/theme.html
@@ -1,12 +1,4 @@
 <template name="theme">
-
-
-
-
-
-
-
-
       <div class="container">
 
         <div class="page-header" id="banner">
