--- views/templates/users/communities.html
@@ -1,68 +1,0 @@
-<template name="communities">
-    {{#if activeBounty}}
-    <div class="alert alert-danger" role="alert">
-    <h5 class="text-center">{{timeRemaining}}</h5><button class="btn btn-danger" id="js-cancel">Cancel</button></div>
-    {{/if}}
-    <div id="outstandingRatings" class="text-center">You still have to answer {{outstandingRatings}} questions.
-    {{#with questions}}
-        <h3 style="margin-top:50px;">{{questionText}}?</h3>
-        <div class="row" style="margin-top:50px;">
-            <div class="col-4 text-center choice" id={{currency0Id}}><h3>{{currency0Name}}<br />
-                <img src="https://raw.githubusercontent.com/reactjs/redux/master/logo/logo.png" width="100px" /></h3>
-            </div>
-            <div class="col-4 text-center choice" id="tie"><br />Both are about the same<br /></div>
-            <div class="col-4 text-center choice" id={{currency1Id}}><h3>{{currency1Name}}<br />
-                <img src="https://raw.githubusercontent.com/reactjs/redux/master/logo/logo.png" width="100px" /></h3>
-            </div>
-        </div>
-    {{/with}}
-    </div>
-
-    <div id="currencyChoices" style="display:none;">
-        <b>You have already added communities for the following cryptocurrencies:</b>
-        <div class="row" style="margin-top: 15px;">
-            {{#if subsCacheReady}}
-                {{#each alreadyAdded}}
-                    <div class="col-sm-4 col-md-4 col-lg-4">
-                        <a href="/currency/{{slug}}">{{currencyName}}</a>
-                    </div>
-                {{else}}
-                    {{> empty}}
-                {{/each}}
-            {{else}}
-                {{> loading}}
-            {{/if}}
-        </div>
-        <br />
-        <b>Please select any currencies that you actually use (and you're part of its community) and you haven't added previously.</b> <br />
-        If you don't see your currency, you might want to <a href="/addcoin">add it</a>. <br />
-        <br />
-        <b>Filter available cryptocurrencies:</b>
-        <form id="js-form">
-            Name: <input type="text" id="js-name" style="width: 50%" placeholder="Enter the name of the cryptocurrency" /><br />
-            Symbol: <input type="text" id="js-symbol" style="width: 50%" placeholder="Enter the symbol of the cryptocurrency" />
-        </form>
-        <div class="row" style="margin-top: 15px;">
-            {{#if subsCacheReady}}
-                {{#each currencies}}
-                    <div class="col-sm-4 col-md-4 col-lg-4">
-                        <a href="#" class="toggle">{{currencyName}}</a>
-                        <div class="row" style="display:none;" id="links_{{_id}}">
-                            <p>Please add a URL to a {{currencyName}} community you are a part of:<br /> 
-                            <input type='text' id='js-com-url_{{_id}}' placeholder="URL here"><button class="js-save">Save</button></p>
-                        </div>
-                    </div>
-                {{else}}
-                    {{> empty}}
-                {{/each}}
-            {{else}}
-                {{> loading}}
-            {{/if}}
-        </div>
-        <button id="populateRatings">Next</button>
-    </div>
-    <hr />
-    <h5>ADMIN SHORTCUTS</h5>
-    <button id="elo" class="btn btn-danger">ELO</button>
-    <button id="communities" class="btn btn-danger">COMMUNITIES</button>
-</template>
