--- views/templates/users/developers.html
@@ -1,26 +1,0 @@
-<template name="developers">
-	{{#if isDeveloper}}
-		{{#if subsCacheReady}}
-			{{#each developers}}
-				<div class="row" style="display:{{display}};border:solid;">
-					<div class="col-12" style="display:block;">
-						{{username}} has submitted the following profiles: <br />
-						{{#each proofs}}
-						{{service}}: <a href="{{profile}}">{{profile}}</a><br />
-						{{/each}}
-						<br />
-						<button class="btn btn-success" id="approve">APPROVE <i class="fa fa-check" aria-hidden="true"></i></button>
-						<button class="btn btn-danger" id="reject">REJECT <i class="fa fa-times" aria-hidden="true"></i></button>
-						<br />
-					</div>
-				</div>
-			{{else}}
-				All good. No developers to approve.
-			{{/each}}
-		{{else}}
-			{{> loading}}
-		{{/if}}
-	{{else}}
-		You have to be a developer in order to review other developers.
-	{{/if}}
-</template>
